feat: add webhook to check subnet deletion.
